requirements.txt
deta==1.1.0
fastapi==0.95.0
requests==2.28.2
uvicorn==0.21.1
openai
python-dotenv